third_party/WebKit/LayoutTests/editing/selection/drag-drop-events-expected.txt

 
This test verifies the order of events fired by drag-and-drop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
focus src
drop dst
input src
change src
blur src
focus dst
input dst
